# backend/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: FleetSpecial API
  version: 1.0.0
  description: |
    Contrato **REST (API First)** de FleetSpecial — SaaS multi-tenant para transporte
    especial y flotas pequeñas en Colombia.

    Este contrato **antecede al código** (Spec Driven Development): de él se generan el
    SDK del portal web, los mocks para que web y móvil avancen en paralelo, y la
    validación de requests en runtime. Deriva de las specs aprobadas (Fase 3) y del
    modelo de dominio (Fase 2).

    ## Trazabilidad spec → endpoints
    - spec-001 Onboarding de Empresa (Tenant) ....... POST /tenants
    - spec-002 Invitar usuarios y roles ............. POST /usuarios, /usuarios/{id}
    - spec-003 Registrar Vehículo ................... /vehiculos
    - spec-004 Registrar Conductor + Licencia ....... /conductores
    - spec-005 Registrar Documento + adjunto ........ /documentos
    - spec-006 Alertas / Semáforo ................... GET /cumplimiento/*
    - spec-007 Renovación de Documento .............. POST /documentos/{id}/renovaciones
    - spec-008 Crear Servicio + Asignación .......... /servicios, /asignaciones
    - spec-009 Regla de oro (bloqueo asignación) .... efecto en POST /asignaciones (409)
    - spec-010 Conductor offline (mi día, sync) ..... /sync/*, /servicios/{id}/estado
    - spec-011 Tanqueo offline append-only .......... POST /combustible (Idempotency-Key)

    ## Convenciones (Fase 5)
    - Versionado por **prefijo de ruta** (`/v1`).
...
  contact:
...
  license:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /servicios:
    get:
      tags: [Servicios]
      summary: Listar servicios (agenda)
      operationId: listarServicios
      parameters:
        - name: desde
          in: query
          schema: { type: string, format: date-time }
        - name: hasta
          in: query
          schema: { type: string, format: date-time }
        - name: estado
          in: query
          schema: { $ref: '#/components/schemas/EstadoServicio' }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Página de servicios.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Pagina'
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/Servicio' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Servicios]
      summary: Crear un servicio
      description: spec-008. Origen, destino, ventana horaria y cliente.
      operationId: crearServicio
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CrearServicioRequest' }
      responses:
        '201':
          description: Servicio creado (estado Planificado).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Servicio' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /servicios/{servicioId}/asignacion:
    parameters:
      - $ref: '#/components/parameters/ServicioId'
    put:
      tags: [Servicios]
      summary: Asignar vehículo + conductor a un servicio
      description: |
        spec-008 + spec-009 (**regla de oro**). Valida que no haya **choque de ventana
        horaria** (409 `conflicto_horario`) y que el vehículo y el conductor estén
        **al día documentalmente** (consulta ACL a Cumplimiento):
        - Semáforo **rojo** ⇒ se rechaza con `409 incumplimiento` (emite `AsignacionRechazada`).
        - Semáforo **amarillo** ⇒ se permite pero la respuesta incluye `advertencias`.
        En éxito emite `ServicioAsignado`.
      operationId: asignarServicio
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AsignarServicioRequest' }
      responses:
        '200':
          description: Asignación creada (puede incluir advertencias por semáforo amarillo).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Asignacion' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: |
            Conflicto. `type` indica el motivo: `conflicto_horario` (solapamiento) o
            `incumplimiento` (semáforo rojo, regla de oro — spec-009).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /servicios/{servicioId}/estado:
    parameters:
      - $ref: '#/components/parameters/ServicioId'
    post:
      tags: [Servicios]
      summary: Cambiar el estado del servicio (iniciar/finalizar/cancelar)
      description: |
        spec-010. Transiciones válidas: `Planificado→Iniciado→Finalizado` o
        `Planificado→Cancelado` (invariante S2). El conductor puede ejecutarlo **offline**
        y sincronizarlo después; por eso acepta `Idempotency-Key` y marca de tiempo del
        cliente. Emite `ServicioIniciado` / `ServicioFinalizado`.
      operationId: cambiarEstadoServicio
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CambiarEstadoServicioRequest' }
      responses:
        '200':
          description: Estado actualizado.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Servicio' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Transición de estado inválida (viola invariante S2).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT OIDC. El token incluye el claim `tenant_id` y los `roles`. El servidor
        deriva el tenant del token (nunca del body/query) y lo fija en la sesión de
        base de datos para que Row Level Security aísle los datos (ADR-0008).
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: UUID generado por el cliente para deduplicar reintentos (offline).
      schema: { type: string, format: uuid }
...
    ServicioId:
      name: servicioId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Unauthorized:
      description: Falta o es inválido el token de autenticación.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    NotFound:
      description: Recurso no encontrado en el tenant actual.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ValidationError:
      description: Error de validación de la solicitud.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Problem:
      type: object
      description: Error con forma RFC 7807.
      properties:
        type: { type: string, description: 'Identificador del tipo de error (p. ej. conflicto_horario, incumplimiento).' }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: array
          description: Detalle de errores de validación por campo.
          items:
            type: object
            properties:
              campo: { type: string }
              mensaje: { type: string }
      required: [title, status]
    Pagina:
      type: object
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
      required: [page, pageSize, total]
...
    EstadoServicio:
      type: string
      enum: [Planificado, Iniciado, Finalizado, Cancelado]
    VentanaHoraria:
      type: object
      properties:
        inicio: { type: string, format: date-time }
        fin: { type: string, format: date-time }
      required: [inicio, fin]
    CrearServicioRequest:
      type: object
      properties:
        origen: { type: string }
        destino: { type: string }
        ventana: { $ref: '#/components/schemas/VentanaHoraria' }
        cliente: { type: string }
      required: [origen, destino, ventana]
    AsignarServicioRequest:
      type: object
      properties:
        vehiculoId: { type: string, format: uuid }
        conductorId: { type: string, format: uuid }
      required: [vehiculoId, conductorId]
    CambiarEstadoServicioRequest:
      type: object
      properties:
        accion: { type: string, enum: [iniciar, finalizar, cancelar] }
        ocurridoEn: { type: string, format: date-time, description: 'Marca de tiempo del cliente (puede ser offline).' }
        odometro: { type: integer, description: 'Lectura al iniciar/finalizar, si aplica.' }
        clientId: { type: string, format: uuid, description: 'UUID del cambio para idempotencia.' }
      required: [accion]
    Asignacion:
      type: object
      properties:
        servicioId: { type: string, format: uuid }
        vehiculoId: { type: string, format: uuid }
        conductorId: { type: string, format: uuid }
        advertencias:
          type: array
          description: 'Advertencias no bloqueantes (p. ej. semáforo amarillo — spec-009).'
          items: { type: string }
      required: [servicioId, vehiculoId, conductorId]
    Servicio:
      type: object
      properties:
        id: { type: string, format: uuid }
        origen: { type: string }
        destino: { type: string }
        ventana: { $ref: '#/components/schemas/VentanaHoraria' }
        cliente: { type: string }
        estado: { $ref: '#/components/schemas/EstadoServicio' }
        asignacion: { $ref: '#/components/schemas/Asignacion' }
        inicioReal: { type: string, format: date-time }
        finReal: { type: string, format: date-time }
      required: [id, origen, destino, ventana, estado]
...